Added "show only to these groups" feature to custom buttons - there's one high priority bug still with this new feature: saved group lists do not show up when re-editing a button...select the group list again (screen_list_custom_save.php, screen.php, screen_list_custom_sections.html, entriesdisplay.php) Fixed a bug in the inclusion of language files at the top of functions.php Added support for showing the main form on screen when editing a subform entry, in a many-to-one relationship - previously only the subform entry would show up (functions.php) Added new notification option - Send message to the e-mail address specified in element X (functions.php) Fixed bug with the new createProceduresTable feature, which caused Procedures to fail if they did not use that function - it was trying to drop tables that you hadn't created (advancedCalculation.php)
git-svn-id: svn+ssh://svn.code.sf.net/p/formulize/code@475 e5624274-254a-0410-876d-c0e85fc73166

+++ Formulize/trunk/class/advancedCalculation.php
@@ -258,9 +258,11 @@ class formulizeAdvancedCalculationHandler {
   // this function removes temp tables created by the createProceduresTable on this pageload
   function destroyTables() {
     global $xoopsDB;
-    $sql = "DROP TABLE `".implode("`, `",$GLOBALS['formulize_procedures_tablenames'])."`;";
-    if(!$res = $xoopsDB->query($sql)) {
-	print "Error: could not drop the temporary tables created by this procedure.<br>".mysql_error()."<br>$sql";
+    if(isset($GLOBALS['formulize_procedures_tablenames'])) {
+        $sql = "DROP TABLE `".implode("`, `",$GLOBALS['formulize_procedures_tablenames'])."`;";
+        if(!$res = $xoopsDB->query($sql)) {
+        	print "Error: could not drop the temporary tables created by this procedure.<br>".mysql_error()."<br>$sql";
+        }
     }
   }
